Изменил структуру директории компонентов
