+++ src/components/favorites-components/favorites-card.tsx
@@ -1,37 +1,32 @@
-import { Link } from 'react-router-dom';
-import { Offer } from '../../../types';
-import { AppRoutes } from '../../../const';
+import {Offer} from '../../types';
 import cn from 'classnames';
 
-type OfferCardProps = {
+type FavoritesCardProps = {
   offer: Offer;
 }
 
-function OfferCard({offer} : OfferCardProps) : JSX.Element {
+function FavoritesCard({offer} : FavoritesCardProps) : JSX.Element {
   return (
-    <article className="cities__card place-card">
+    <article className="favorites__card place-card">
       <div className={cn('place-card__mark', {'visually-hidden': !offer.isPremium})}>
         <span>Premium</span>
       </div>
-      <div className="cities__image-wrapper place-card__image-wrapper">
-        <Link to={`${AppRoutes.Offer}/${offer.id}`}>
-          <img className="place-card__image" src={offer.previewImage} width="260" height="200" alt="Place image"/>
-        </Link>
+      <div className="favorites__image-wrapper place-card__image-wrapper">
+        <a href="#">
+          <img className="place-card__image" src={offer.previewImage} width="150" height="110" alt="Place image"/>
+        </a>
       </div>
-      <div className="place-card__info">
+      <div className="favorites__card-info place-card__info">
         <div className="place-card__price-wrapper">
           <div className="place-card__price">
             <b className="place-card__price-value">&euro;{offer.price}</b>
             <span className="place-card__price-text">&#47;&nbsp;night</span>
           </div>
-          <button className={
-            cn('place-card__bookmark-button', 'button', {'place-card__bookmark-button--active': offer.isFavorite})
-          } type="button"
-          >
+          <button className="place-card__bookmark-button place-card__bookmark-button--active button" type="button">
             <svg className="place-card__bookmark-icon" width="18" height="19">
               <use xlinkHref="#icon-bookmark"></use>
             </svg>
-            <span className="visually-hidden">To bookmarks</span>
+            <span className="visually-hidden">In bookmarks</span>
           </button>
         </div>
         <div className="place-card__rating rating">
@@ -45,7 +40,8 @@ function OfferCard({offer} : OfferCardProps) : JSX.Element {
         </h2>
         <p className="place-card__type">{offer.type}</p>
       </div>
-    </article>);
+    </article>
+  );
 }
 
-export default OfferCard;
+export default FavoritesCard;
