+++ src/components/favorites-components/favorites-list-item.tsx
