+++ src/components/main-components/main-offers-empty.tsx
