+++ src/components/offer-components/offer-image.tsx
