+++ src/components/offer-components/offer-inside-item.tsx
