+++ src/pages/login-page/login-page.tsx
@@ -1,7 +1,6 @@
-import { Link } from 'react-router-dom';
-import Logo from '../../components/logo/logo';
-import { AppRoutes, Cities } from '../../const';
 import { Helmet } from 'react-helmet-async';
+import HeaderContainer from '../../components/header/header-container';
+import LoginMain from '../../components/login-components/login-main';
 
 function LoginPage() : JSX.Element {
   return (
@@ -9,41 +8,9 @@ function LoginPage() : JSX.Element {
       <Helmet>
         <title>6 cities - Login</title>
       </Helmet>
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Logo/>
-            </div>
-          </div>
-        </div>
-      </header>
+      <HeaderContainer isNavShown={false}/>
 
-      <main className="page__main page__main--login">
-        <div className="page__login-container container">
-          <section className="login">
-            <h1 className="login__title">Sign in</h1>
-            <form className="login__form form" action="#" method="post">
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">E-mail</label>
-                <input className="login__input form__input" type="email" name="email" placeholder="Email" required/>
-              </div>
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">Password</label>
-                <input className="login__input form__input" type="password" name="password" placeholder="Password" required/>
-              </div>
-              <button className="login__submit form__submit button" type="submit">Sign in</button>
-            </form>
-          </section>
-          <section className="locations locations--login locations--current">
-            <div className="locations__item">
-              <Link className="locations__item-link" to={AppRoutes.Main} state={{ city: Cities.Amsterdam}}>
-                <span>Amsterdam</span>
-              </Link>
-            </div>
-          </section>
-        </div>
-      </main>
+      <LoginMain/>
     </div>
   );
 }
