+++ src/pages/main-page/main.tsx
@@ -1,14 +1,10 @@
-import Logo from '../../components/logo/logo';
 import { Cities } from '../../const';
-import LocationItem from '../../components/locations-item/location-item';
 import { Offers } from '../../types';
 import React from 'react';
 import { useLocation } from 'react-router-dom';
-import MainOffersEmpty from '../../components/main-components/main-offers-empty/main-offers-empty';
-import MainOffers from '../../components/main-components/main-offers/main-offers';
 import { Helmet } from 'react-helmet-async';
-import ProfileNavLink from '../../components/header-nav-links/profile-nav-link/profile-nav-link';
-import SignoutLink from '../../components/header-nav-links/signout-link/signout-link';
+import HeaderContainer from '../../components/header/header-container';
+import MainMain from '../../components/main-components/main-main';
 
 type MainPageProps = {
   offers: Offers;
@@ -28,43 +24,9 @@ function MainPage({offers, city = Cities.Paris} : MainPageProps) : JSX.Element {
       <Helmet>
         <title>6 cities</title>
       </Helmet>
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Logo/>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <ProfileNavLink/>
-                <SignoutLink/>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
+      <HeaderContainer isNavShown/>
 
-      <main className="page__main page__main--index">
-        <h1 className="visually-hidden">Cities</h1>
-        <div className="tabs">
-          <section className="locations container">
-            <ul className="locations__list tabs__list">
-              {Object.values(Cities).map((c) => (
-                <LocationItem
-                  key={crypto.randomUUID()}
-                  city={c}
-                  currentCity={currentCity.city}
-                />
-              ))}
-            </ul>
-          </section>
-        </div>
-        <div className="cities">
-          {offersInCurrentCity.length === 0 ?
-            <MainOffersEmpty/> :
-            <MainOffers city={currentCity.city} offersInCurrentCity={offersInCurrentCity}/>}
-        </div>
-      </main>
+      <MainMain offersInCurrentCity={offersInCurrentCity} city={currentCity.city}/>
     </div>);
 }
