+++ src/pages/page-404/page-404.tsx
@@ -1,7 +1,7 @@
 import { Link } from 'react-router-dom';
-import Logo from '../../components/logo/logo';
 import { Helmet } from 'react-helmet-async';
 import { AppRoutes } from '../../const';
+import HeaderContainer from '../../components/header/header-container';
 
 function Page404 () : JSX.Element {
   return (
@@ -9,15 +9,8 @@ function Page404 () : JSX.Element {
       <Helmet>
         <title>6 cities - 404</title>
       </Helmet>
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Logo/>
-            </div>
-          </div>
-        </div>
-      </header>
+      <HeaderContainer isNavShown={false}/>
+
       <main className="page__main page__main--404">
         <div className="page__404-container container">
           <section className="page__404" style={{textAlign: 'center'}}>
