+++ src/util.ts
@@ -1,0 +1,4 @@
+const capitalizeFirstLetter = (str: string) => str.charAt(0).toUpperCase() + str.slice(1);
+
+
+export { capitalizeFirstLetter };
